process status updates in separate thread
