+++ SingularityService/src/main/java/com/hubspot/singularity/SingularityMainModule.java
@@ -56,6 +56,7 @@ import com.hubspot.singularity.hooks.LoadBalancerClient;
 import com.hubspot.singularity.hooks.LoadBalancerClientImpl;
 import com.hubspot.singularity.hooks.SingularityWebhookPoller;
 import com.hubspot.singularity.hooks.SingularityWebhookSender;
+import com.hubspot.singularity.mesos.SingularityMesosStatusUpdateHandler;
 import com.hubspot.singularity.metrics.SingularityGraphiteReporterManaged;
 import com.hubspot.singularity.sentry.NotifyingExceptionMapper;
 import com.hubspot.singularity.sentry.SingularityExceptionNotifier;
@@ -160,6 +161,8 @@ public class SingularityMainModule implements Module {
         "check-new-task")).in(Scopes.SINGLETON);
 
     binder.bind(SingularityGraphiteReporterManaged.class).in(Scopes.SINGLETON);
+
+    binder.bind(SingularityMesosStatusUpdateHandler.class).in(Scopes.SINGLETON);
   }
 
   @Provides
