+++ SingularityService/src/main/java/com/hubspot/singularity/config/SingularityConfiguration.java
@@ -272,6 +272,11 @@ public class SingularityConfiguration extends Configuration {
   @Max(5)
   private double schedulerPriorityWeightFactor = 1.0;
 
+  @Min(1)
+  private int statusUpdateQueueCapacity = 10;
+
+  private boolean processStatusUpdatesInSeparateThread = false;
+
   public long getAskDriverToKillTasksAgainAfterMillis() {
     return askDriverToKillTasksAgainAfterMillis;
   }
@@ -1086,5 +1091,21 @@ public class SingularityConfiguration extends Configuration {
 
   public void setSchedulerPriorityWeightFactor(double schedulerPriorityWeightFactor) {
     this.schedulerPriorityWeightFactor = schedulerPriorityWeightFactor;
+  }
+
+  public int getStatusUpdateQueueCapacity() {
+    return statusUpdateQueueCapacity;
+  }
+
+  public void setStatusUpdateQueueCapacity(int statusUpdateQueueCapacity) {
+    this.statusUpdateQueueCapacity = statusUpdateQueueCapacity;
+  }
+
+  public boolean isProcessStatusUpdatesInSeparateThread() {
+    return processStatusUpdatesInSeparateThread;
+  }
+
+  public void setProcessStatusUpdatesInSeparateThread(boolean processStatusUpdatesInSeparateThread) {
+    this.processStatusUpdatesInSeparateThread = processStatusUpdatesInSeparateThread;
   }
 }
