+++ SingularityService/src/main/java/com/hubspot/singularity/mesos/SingularityMesosScheduler.java
@@ -11,7 +11,6 @@ import javax.inject.Singleton;
 
 import org.apache.mesos.Protos;
 import org.apache.mesos.Protos.Offer;
-import org.apache.mesos.Protos.TaskStatus.Reason;
 import org.apache.mesos.Scheduler;
 import org.apache.mesos.SchedulerDriver;
 import org.slf4j.Logger;
@@ -19,43 +18,25 @@ import org.slf4j.LoggerFactory;
 
 import com.codahale.metrics.annotation.Timed;
 import com.google.common.base.Optional;
-import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import com.google.inject.name.Named;
 import com.hubspot.mesos.JavaUtils;
 import com.hubspot.mesos.MesosUtils;
 import com.hubspot.mesos.Resources;
-import com.hubspot.singularity.ExtendedTaskState;
-import com.hubspot.singularity.InvalidSingularityTaskIdException;
-import com.hubspot.singularity.SingularityCreateResult;
-import com.hubspot.singularity.SingularityMainModule;
-import com.hubspot.singularity.SingularityPendingDeploy;
 import com.hubspot.singularity.SingularityPriorityFreezeParent;
-import com.hubspot.singularity.SingularityRequestWithState;
 import com.hubspot.singularity.SingularityTask;
-import com.hubspot.singularity.SingularityTaskHistoryUpdate;
-import com.hubspot.singularity.SingularityTaskId;
 import com.hubspot.singularity.SingularityTaskRequest;
 import com.hubspot.singularity.SingularityTaskRequestWithPriority;
-import com.hubspot.singularity.SingularityTaskStatusHolder;
 import com.hubspot.singularity.SlaveMatchState;
 import com.hubspot.singularity.config.CustomExecutorConfiguration;
 import com.hubspot.singularity.config.MesosConfiguration;
 import com.hubspot.singularity.config.SingularityConfiguration;
-import com.hubspot.singularity.data.DeployManager;
 import com.hubspot.singularity.data.PriorityManager;
-import com.hubspot.singularity.data.RequestManager;
 import com.hubspot.singularity.data.TaskManager;
-import com.hubspot.singularity.data.transcoders.IdTranscoder;
-import com.hubspot.singularity.data.transcoders.SingularityTranscoderException;
-import com.hubspot.singularity.scheduler.SingularityHealthchecker;
-import com.hubspot.singularity.scheduler.SingularityNewTaskChecker;
 import com.hubspot.singularity.scheduler.SingularityScheduler;
 import com.hubspot.singularity.scheduler.SingularitySchedulerStateCache;
-import com.hubspot.singularity.sentry.SingularityExceptionNotifier;
 
 @Singleton
 public class SingularityMesosScheduler implements Scheduler {
@@ -65,52 +46,39 @@ public class SingularityMesosScheduler implements Scheduler {
   private final Resources defaultResources;
   private final Resources defaultCustomExecutorResources;
   private final TaskManager taskManager;
-  private final RequestManager requestManager;
-  private final DeployManager deployManager;
   private final PriorityManager priorityManager;
   private final SingularityScheduler scheduler;
   private final SingularityConfiguration configuration;
   private final SingularityMesosTaskBuilder mesosTaskBuilder;
   private final SingularityMesosFrameworkMessageHandler messageHandler;
-  private final SingularityHealthchecker healthchecker;
-  private final SingularityNewTaskChecker newTaskChecker;
   private final SingularitySlaveAndRackManager slaveAndRackManager;
-  private final SingularityLogSupport logSupport;
   private final SingularityTaskSizeOptimizer taskSizeOptimizer;
 
-  private final SingularityExceptionNotifier exceptionNotifier;
 
   private final Provider<SingularitySchedulerStateCache> stateCacheProvider;
-  private final String serverId;
   private final SchedulerDriverSupplier schedulerDriverSupplier;
 
-  private final IdTranscoder<SingularityTaskId> taskIdTranscoder;
+  private final SingularityMesosStatusUpdateHandler statusUpdateHandler;
 
   @Inject
-  SingularityMesosScheduler(MesosConfiguration mesosConfiguration, SingularityConfiguration configuration, TaskManager taskManager, PriorityManager priorityManager, SingularityScheduler scheduler, SingularitySlaveAndRackManager slaveAndRackManager,
-      SingularityNewTaskChecker newTaskChecker, SingularityMesosTaskBuilder mesosTaskBuilder, SingularityLogSupport logSupport, RequestManager requestManager,
-      Provider<SingularitySchedulerStateCache> stateCacheProvider, SingularityHealthchecker healthchecker, DeployManager deployManager, SingularityExceptionNotifier exceptionNotifier,SingularityMesosFrameworkMessageHandler messageHandler,
-      @Named(SingularityMainModule.SERVER_ID_PROPERTY) String serverId, SchedulerDriverSupplier schedulerDriverSupplier, SingularityTaskSizeOptimizer taskSizeOptimizer, final IdTranscoder<SingularityTaskId> taskIdTranscoder, CustomExecutorConfiguration customExecutorConfiguration) {
+  public SingularityMesosScheduler(MesosConfiguration mesosConfiguration, CustomExecutorConfiguration customExecutorConfiguration, TaskManager taskManager, PriorityManager priorityManager,
+      SingularityScheduler scheduler, SingularityConfiguration configuration, SingularityMesosTaskBuilder mesosTaskBuilder,
+      SingularityMesosFrameworkMessageHandler messageHandler, SingularitySlaveAndRackManager slaveAndRackManager, SingularityTaskSizeOptimizer taskSizeOptimizer,
+      Provider<SingularitySchedulerStateCache> stateCacheProvider, SchedulerDriverSupplier schedulerDriverSupplier, SingularityMesosStatusUpdateHandler statusUpdateHandler) {
     this.defaultResources = new Resources(mesosConfiguration.getDefaultCpus(), mesosConfiguration.getDefaultMemory(), 0, mesosConfiguration.getDefaultDisk());
     this.defaultCustomExecutorResources = new Resources(customExecutorConfiguration.getNumCpus(), customExecutorConfiguration.getMemoryMb(), 0, customExecutorConfiguration.getDiskMb());
     this.taskManager = taskManager;
-    this.deployManager = deployManager;
     this.priorityManager = priorityManager;
-    this.newTaskChecker = newTaskChecker;
-    this.slaveAndRackManager = slaveAndRackManager;
     this.scheduler = scheduler;
-    this.messageHandler = messageHandler;
-    this.taskSizeOptimizer = taskSizeOptimizer;
-    this.mesosTaskBuilder = mesosTaskBuilder;
-    this.logSupport = logSupport;
-    this.stateCacheProvider = stateCacheProvider;
-    this.healthchecker = healthchecker;
-    this.serverId = serverId;
-    this.schedulerDriverSupplier = schedulerDriverSupplier;
-    this.taskIdTranscoder = taskIdTranscoder;
-    this.exceptionNotifier = exceptionNotifier;
-    this.requestManager = requestManager;
     this.configuration = configuration;
+    this.mesosTaskBuilder = mesosTaskBuilder;
+    this.messageHandler = messageHandler;
+    this.slaveAndRackManager = slaveAndRackManager;
+    this.taskSizeOptimizer = taskSizeOptimizer;
+    this.stateCacheProvider = stateCacheProvider;
+    this.schedulerDriverSupplier = schedulerDriverSupplier;
+
+    this.statusUpdateHandler = statusUpdateHandler;
   }
 
   @Override
@@ -308,141 +276,9 @@ public class SingularityMesosScheduler implements Scheduler {
     LOG.info("Offer {} rescinded", offerId);
   }
 
-  /**
-   * 1- we have a previous update, and this is a duplicate of it (ignore) 2- we don't have a
-   * previous update, 2 cases: a - this task has already been destroyed (we can ignore it then) b -
-   * we've never heard of this task (very unlikely since we first write a status into zk before we
-   * launch a task)
-   */
-  private boolean isDuplicateOrIgnorableStatusUpdate(Optional<SingularityTaskStatusHolder> previousTaskStatusHolder, final SingularityTaskStatusHolder newTaskStatusHolder) {
-    if (!previousTaskStatusHolder.isPresent()) {
-      return true;
-    }
-
-    if (!previousTaskStatusHolder.get().getTaskStatus().isPresent()) { // this is our launch state
-      return false;
-    }
-
-    return previousTaskStatusHolder.get().getTaskStatus().get().getState() == newTaskStatusHolder.getTaskStatus().get().getState();
-  }
-
-  private void saveNewTaskStatusHolder(SingularityTaskId taskIdObj, SingularityTaskStatusHolder newTaskStatusHolder, ExtendedTaskState taskState) {
-    if (taskState.isDone()) {
-      taskManager.deleteLastActiveTaskStatus(taskIdObj);
-    } else {
-      taskManager.saveLastActiveTaskStatus(newTaskStatusHolder);
-    }
-  }
-
-  private Optional<SingularityTaskId> getTaskId(String taskId) {
-    try {
-      return Optional.of(taskIdTranscoder.fromString(taskId));
-    } catch (InvalidSingularityTaskIdException | SingularityTranscoderException e) {
-      exceptionNotifier.notify(e);
-      LOG.error("Unexpected taskId {} ", taskId, e);
-      return Optional.absent();
-    }
-  }
-
-  private Optional<String> getStatusMessage(Protos.TaskStatus status, Optional<SingularityTask> task) {
-    if (status.hasMessage() && !Strings.isNullOrEmpty(status.getMessage())) {
-      return Optional.of(status.getMessage());
-    } else if (status.hasReason() && status.getReason() == Reason.REASON_CONTAINER_LIMITATION_MEMORY) {
-      if (task.isPresent() && task.get().getTaskRequest().getDeploy().getResources().isPresent()) {
-        if (task.get().getTaskRequest().getDeploy().getResources().get().getDiskMb() > 0) {
-          return Optional.of(String.format("Task exceeded one or more memory limits (%s MB mem, %s MB disk).", task.get().getTaskRequest().getDeploy().getResources().get().getMemoryMb(), task.get().getTaskRequest().getDeploy().getResources().get().getDiskMb()));
-        } else {
-          return Optional.of(String.format("Task exceeded memory limit (%s MB mem).", task.get().getTaskRequest().getDeploy().getResources().get().getMemoryMb()));
-        }
-
-      }
-      return Optional.of("Task exceeded memory limit");
-    }
-
-    return Optional.absent();
-  }
-
   @Override
-  @Timed
   public void statusUpdate(SchedulerDriver driver, Protos.TaskStatus status) {
-    final String taskId = status.getTaskId().getValue();
-
-    long timestamp = System.currentTimeMillis();
-
-    if (status.hasTimestamp()) {
-      timestamp = (long) (status.getTimestamp() * 1000);
-    }
-
-    LOG.debug("Task {} is now {} ({}) at {} ", taskId, status.getState(), status.getMessage(), timestamp);
-
-    final Optional<SingularityTaskId> maybeTaskId = getTaskId(taskId);
-
-    if (!maybeTaskId.isPresent()) {
-      return;
-    }
-
-    final SingularityTaskId taskIdObj = maybeTaskId.get();
-
-    final SingularityTaskStatusHolder newTaskStatusHolder = new SingularityTaskStatusHolder(taskIdObj, Optional.of(status), System.currentTimeMillis(), serverId, Optional.<String>absent());
-    final Optional<SingularityTaskStatusHolder> previousTaskStatusHolder = taskManager.getLastActiveTaskStatus(taskIdObj);
-    final ExtendedTaskState taskState = ExtendedTaskState.fromTaskState(status.getState());
-
-    if (isDuplicateOrIgnorableStatusUpdate(previousTaskStatusHolder, newTaskStatusHolder)) {
-      LOG.trace("Ignoring status update {} to {}", taskState, taskIdObj);
-      saveNewTaskStatusHolder(taskIdObj, newTaskStatusHolder, taskState);
-      return;
-    }
-
-    final Optional<SingularityTask> task = taskManager.getTask(taskIdObj);
-
-    final boolean isActiveTask = taskManager.isActiveTask(taskId);
-
-    if (isActiveTask && !taskState.isDone()) {
-      if (task.isPresent()) {
-        final Optional<SingularityPendingDeploy> pendingDeploy = deployManager.getPendingDeploy(taskIdObj.getRequestId());
-
-        Optional<SingularityRequestWithState> requestWithState = Optional.absent();
-
-        if (taskState == ExtendedTaskState.TASK_RUNNING) {
-          requestWithState = requestManager.getRequest(taskIdObj.getRequestId());
-          healthchecker.enqueueHealthcheck(task.get(), pendingDeploy, requestWithState);
-        }
-
-        if (!pendingDeploy.isPresent() || !pendingDeploy.get().getDeployMarker().getDeployId().equals(taskIdObj.getDeployId())) {
-          if (!requestWithState.isPresent()) {
-            requestWithState = requestManager.getRequest(taskIdObj.getRequestId());
-          }
-          newTaskChecker.enqueueNewTaskCheck(task.get(), requestWithState, healthchecker);
-        }
-      } else {
-        final String message = String.format("Task %s is active but is missing task data", taskId);
-        exceptionNotifier.notify(message);
-        LOG.error(message);
-      }
-    }
-
-    final Optional<String> statusMessage = getStatusMessage(status, task);
-
-    final SingularityTaskHistoryUpdate taskUpdate =
-        new SingularityTaskHistoryUpdate(taskIdObj, timestamp, taskState, statusMessage, status.hasReason() ? Optional.of(status.getReason().name()) : Optional.<String>absent());
-    final SingularityCreateResult taskHistoryUpdateCreateResult = taskManager.saveTaskHistoryUpdate(taskUpdate);
-
-    logSupport.checkDirectory(taskIdObj);
-
-    if (taskState.isDone()) {
-      healthchecker.cancelHealthcheck(taskId);
-      newTaskChecker.cancelNewTaskCheck(taskId);
-
-      taskManager.deleteKilledRecord(taskIdObj);
-
-      SingularitySchedulerStateCache stateCache = stateCacheProvider.get();
-
-      slaveAndRackManager.checkStateAfterFinishedTask(taskIdObj, status.getSlaveId().getValue(), stateCache);
-
-      scheduler.handleCompletedTask(task, taskIdObj, isActiveTask, timestamp, taskState, taskHistoryUpdateCreateResult, stateCache, status);
-    }
-
-    saveNewTaskStatusHolder(taskIdObj, newTaskStatusHolder, taskState);
+    statusUpdateHandler.enqueueStatusUpdate(status);
   }
 
   @Override
